+++ src/ledger/ledgersignliquidtransactionactivity.cpp
@@ -5,8 +5,6 @@
 
 #include <wally_wrapper.h>
 
-//#define BLINDING_FACTOR_LEN 32 /** Length of a Blinding Factor (or blinder) */
-
 namespace {
 
 QByteArray inputLiquidBytes(const QJsonObject& input, bool strict)
@@ -25,9 +23,11 @@ QByteArray inputLiquidBytes(const QJsonObject& input, bool strict)
         const auto commitment = ParseByteArray(input.value("commitment"));
         stream.writeRawData(commitment.data(), commitment.size());
     } else {
-        stream << uint8_t(0x08);
-        for (int i = 0; i < 32; ++i) {
-            stream << uint8_t(0x00);
+        for (int j = 0; j < 3; j++) {
+            stream << uint8_t(0x08);
+            for (int i = 0; i < 10; ++i) {
+                stream << uint8_t(0x00);
+            }
         }
     }
     return data;
@@ -38,7 +38,6 @@ QByteArray sequenceBytes(const QJsonObject& input)
     QByteArray data;
     QDataStream stream(&data, QIODevice::WriteOnly);
     stream.setByteOrder(QDataStream::LittleEndian);
-    // Q_ASSERT(input.contains("sequence"));
     stream << ParseSequence(input.value("sequence"));
     return data;
 }
@@ -48,11 +47,16 @@ QByteArray sequenceBytes(const QJsonObject& input)
 LedgerSignLiquidTransactionActivity::LedgerSignLiquidTransactionActivity(const QByteArray& transaction, const QJsonArray& signing_inputs, const QJsonArray& outputs, LedgerDevice* device)
     : SignLiquidTransactionActivity(device)
     , m_device(device)
-    , m_transaction(transaction)
     , m_inputs(signing_inputs)
     , m_outputs(outputs)
     , m_batch(new CommandBatch)
 {
+    wally_tx* tx;
+    wally_tx_from_bytes((const unsigned char*) transaction.constData(), transaction.size(), WALLY_TX_FLAG_USE_ELEMENTS, &tx);
+    m_version = tx->version;
+    m_locktime = tx->locktime;
+    wally_tx_free(tx);
+
     connect(m_batch, &Command::error, this, [this] { fail(); });
 }
 
@@ -75,8 +79,6 @@ void LedgerSignLiquidTransactionActivity::exec()
 
     startUntrustedTransaction(true, 0, m_hw_inputs, m_hw_sequences, script0);
 
-    qDebug() << Q_FUNC_INFO << "==============";
-    qDebug() << m_inputs;
     for (const auto input : m_inputs) {
         m_values.append(ParseSatoshi(input.toObject().value("satoshi")));
         m_abfs.append(ReverseByteArray(ParseByteArray(input.toObject().value("assetblinder"))));
@@ -95,8 +97,7 @@ void LedgerSignLiquidTransactionActivity::startUntrustedTransaction(bool new_tra
         QByteArray data;
         QDataStream stream(&data, QIODevice::WriteOnly);
         stream.setByteOrder(QDataStream::LittleEndian);
-        uint32_t version = 0; // m_transaction.value("transaction_version").toInteger();
-        stream << version << varint<uint32_t>(inputs.size());
+        stream << m_version << varint<uint32_t>(inputs.size());
         exchange(apdu(BTCHIP_CLA, BTCHIP_INS_HASH_INPUT_START, 0x00, new_transaction ? 0x06 : 0x80, data));
     }
     for (int i = 0; i < inputs.size(); ++i) {
@@ -126,8 +127,6 @@ void LedgerSignLiquidTransactionActivity::getLiquidCommitment(int output_index)
 
     if (!output.contains("blinding_key")) {
         m_commitments.append(QByteArray());
-        m_abfs.append(QByteArray());
-        m_vbfs.append(QByteArray());
         return getLiquidCommitment(output_index + 1);
     }
 
@@ -220,27 +219,22 @@ QList<QPair<QJsonObject, QByteArray>> LedgerSignLiquidTransactionActivity::outpu
 
     for (int i = 0; i < m_outputs.size(); ++i) {
         const auto output = m_outputs.at(i).toObject();
-//        qDebug() << output;
-        const auto script = ParseByteArray(output.value("scriptpubkey"));
-        if (script.size() == 0) {
+        if (!output.contains("blinding_key")) {
             QByteArray data;
             QDataStream stream(&data, QIODevice::WriteOnly);
-            stream.setByteOrder(QDataStream::BigEndian);
             stream << uint8_t(0x01);
             const auto asset_id = ParseByteArray(output.value("asset_id"));
             Q_ASSERT(asset_id.size() == 32);
             for (int j = asset_id.size() - 1; j >= 0; --j) {
                 stream << uint8_t(asset_id.at(j));
             }
-            stream << uint8_t(0x01) << ParseSatoshi(output.value("satoshi"));
+            stream << uint8_t(0x01);
+            stream.setByteOrder(QDataStream::BigEndian);
+            stream << ParseSatoshi(output.value("satoshi"));
             res.append({{}, data});
         } else {
             res.append({{}, m_commitments.at(i).mid(64, 143)});
         }
-
-        qDebug() << output;
-        // Q_UNREACHABLE();
-        // TODO see https://gl.blockstream.io/blockstream/green/green_android/-/merge_requests/654/diffs#c906275cde9f83dfd32304b6d523e4cfabd03462_97_98
 
         if (output.contains("blinding_key")) {
             const auto blinding_key = ParseByteArray(output.value("blinding_key"));
@@ -248,20 +242,21 @@ QList<QPair<QJsonObject, QByteArray>> LedgerSignLiquidTransactionActivity::outpu
             res.append({{}, eph_public_key});
             res.append({{}, blinding_key});
         } else {
-            QByteArray nonce(1, 0);
-            res.append({{}, nonce});
-            res.append({{}, nonce});
+            QByteArray dummy(1, 0);
+            res.append({{}, dummy});
+            res.append({{}, dummy});
         }
 
         {
+            const auto scriptpubkey = ParseByteArray(output.value("scriptpubkey"));
             QByteArray data;
             QDataStream stream(&data, QIODevice::WriteOnly);
             stream.setByteOrder(QDataStream::LittleEndian);
-            if (script.size() == 0) {
+            if (scriptpubkey.size() == 0) {
                 stream << varint<uint32_t>(0);
             } else {
-                stream << varint<uint32_t>(script.size());
-                stream.writeRawData(script.data(), script.size());
+                stream << varint<uint32_t>(scriptpubkey.size());
+                stream.writeRawData(scriptpubkey.data(), scriptpubkey.size());
             }
 
             QJsonObject message = {
@@ -318,8 +313,6 @@ void LedgerSignLiquidTransactionActivity::finalizeLiquidInputFull()
         }
     }
 
-    const uint32_t locktime = 0; // ParseLocktime(m_transaction.value("transaction_locktime"));
-
     for (i = 0; i < m_hw_inputs.size(); ++i) {
         const auto input = m_inputs.at(i).toObject();
         const auto script = ParseByteArray(input.value("prevout_script"));
@@ -334,18 +327,11 @@ void LedgerSignLiquidTransactionActivity::finalizeLiquidInputFull()
             QVector<uint32_t> user_path = ParsePath(input.value("user_path"));
             const auto path = pathToData(user_path);
 
-            stream.writeRawData(path.data(), data.size());
-
-            stream.setByteOrder(QDataStream::BigEndian);
-
-            // TODO android implementation uses empty pin, like:
-            //    stream << uint8_t(0x00);
-            QString pin = "0";
-            auto _pin = pin.toUtf8();
             stream.writeRawData(path.data(), path.size());
-            stream << uint8_t(pin.size());
-            stream.writeRawData(_pin.data(), _pin.size());
-            stream << uint32_t(locktime);
+            stream << uint8_t(0x00);
+            stream.setByteOrder(QDataStream::BigEndian);
+            stream << m_locktime;
+            stream.setByteOrder(QDataStream::LittleEndian);
             stream << uint8_t(/* SIGHASH_ALL */ 1);
 
             auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_HASH_SIGN, 0x00, 0x00, data));
@@ -374,9 +360,6 @@ LedgerGetBlindingFactorsActivity::LedgerGetBlindingFactorsActivity(const QJsonAr
     , m_batch(new CommandBatch)
 {
     connect(m_batch, &Command::finished, this, [=] {
-        qDebug() << "LedgerGetBlindingFactorsActivity FINISHED";
-        qDebug() << m_abfs;
-        qDebug() << m_vbfs;
         finish();
     });
 }
@@ -431,8 +414,8 @@ void LedgerGetBlindingFactorsActivity::exec()
             });
 
             auto cbfv = getBlindingFactor(i, BTCHIP_BLINDING_FACTOR_AMOUNT);
-            connect(cbfa, &Command::finished, this, [=] {
-                m_vbfs[i] = ReverseByteArray(cbfa->m_response);
+            connect(cbfv, &Command::finished, this, [=] {
+                m_vbfs[i] = ReverseByteArray(cbfv->m_response);
             });
         }
     }
@@ -452,7 +435,6 @@ void LedgerGetBlindingFactorsActivity::startUntrustedTransaction(uint32_t versio
     }
     for (int i = 0; i < inputs.size(); ++i) {
         QByteArray script = i == input_index ? redeem_script : QByteArray();
-
         QByteArray data;
         QDataStream stream(&data, QIODevice::WriteOnly);
         stream.setByteOrder(QDataStream::LittleEndian);
@@ -473,248 +455,3 @@ DeviceCommand* LedgerGetBlindingFactorsActivity::getBlindingFactor(uint32_t inde
     stream << index;
     return exchange(apdu(BTCHIP_CLA, BTCHIP_INS_GET_LIQUID_BLINDING_FACTOR, type, 0x00, data));
 }
-
-/*
-void LedgerGetBlindingFactorsActivity::getLiquidCommitment(int output_index)
-{
-    if (output_index == m_outputs.size()) {
-        return finalizeLiquidInputFull();
-    }
-
-    Q_ASSERT(output_index < m_outputs.size());
-    const auto output = m_outputs.at(output_index).toObject();
-    const bool last = (output_index + 2) == m_outputs.size();
-
-    Q_ASSERT(output.contains("scriptpubkey"));
-    if (output.value("scriptpubkey").toString().isEmpty()) {
-        m_commitments.append(QByteArray());
-        m_abfs.append(QByteArray());
-        m_vbfs.append(QByteArray());
-        return getLiquidCommitment(output_index + 1);
-    }
-
-    Q_ASSERT(output.contains("satoshi"));
-    const quint64 value = ParseSatoshi(output.value("satoshi"));
-    m_values.append(value);
-
-    Q_ASSERT(output.contains("asset_id"));
-    const auto asset_id = ParseByteArray(output.value("asset_id"));
-    Q_ASSERT(asset_id.size() == 32);
-
-    if (last) {
-        QByteArray data;
-        QDataStream stream(&data, QIODevice::WriteOnly);
-        stream.setByteOrder(QDataStream::BigEndian);
-        stream << uint32_t(output_index);
-        auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_GET_LIQUID_BLINDING_FACTOR, 0x01, 0x00, data));
-        connect(c, &Command::finished, [this, c, output_index, value, asset_id] {
-            Q_ASSERT(c->m_response.size() == 32);
-            m_abfs.append(c->m_response);
-
-            QByteArray data;
-            QDataStream stream(&data, QIODevice::WriteOnly);
-            stream.setByteOrder(QDataStream::BigEndian);
-
-            stream.writeRawData(asset_id.data(), asset_id.length());
-            stream << quint64(value) << uint32_t(output_index);
-
-            const auto values = m_values.toVector();
-            const auto abf = m_abfs.join();
-            const auto vbf = m_vbfs.join();
-
-            char final_vbf[BLINDING_FACTOR_LEN];
-            int ret = wally_asset_final_vbf(
-                (const uint64_t*) values.data(), values.size(),
-                m_inputs.size(),
-                (const unsigned char*) abf.data(), abf.size(),
-                (const unsigned char*) vbf.data(), vbf.size(),
-                (unsigned char*) final_vbf, BLINDING_FACTOR_LEN);
-            Q_ASSERT(ret == WALLY_OK);
-            m_vbfs.append(QByteArray(final_vbf, BLINDING_FACTOR_LEN));
-            stream.writeRawData(final_vbf, BLINDING_FACTOR_LEN);
-
-            auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_GET_LIQUID_COMMITMENTS, 0x02, 0x00, data));
-            connect(c, &Command::finished, [this, c, output_index] {
-                m_commitments.append(c->m_response);
-                getLiquidCommitment(output_index + 1);
-            });
-        });
-    } else {
-        QByteArray data;
-        QDataStream stream(&data, QIODevice::WriteOnly);
-        stream.setByteOrder(QDataStream::BigEndian);
-
-        stream.writeRawData(asset_id.data(), asset_id.length());
-        stream << quint64(value) << uint32_t(output_index);
-
-        auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_GET_LIQUID_COMMITMENTS, 0x01, 0x00, data));
-        connect(c, &Command::finished, [this, c, output_index] {
-            Q_ASSERT(c->m_response.size() >= 64);
-            m_commitments.append(c->m_response);
-            m_abfs.append(c->m_response.mid(0, 32));
-            m_vbfs.append(c->m_response.mid(32, 32));
-            getLiquidCommitment(output_index + 1);
-        });
-    }
-}
-
-QList<QPair<QJsonObject, QByteArray>> LedgerGetBlindingFactorsActivity::outputLiquidBytes()
-{
-    QList<QPair<QJsonObject,QByteArray>> res;
-    {
-        QByteArray data;
-        QDataStream stream(&data, QIODevice::WriteOnly);
-        stream.setByteOrder(QDataStream::LittleEndian);
-        stream << varint<uint32_t>(m_outputs.size());
-        res.append({{}, data});
-    }
-
-    for (int i = 0; i < m_outputs.size(); ++i) {
-        const auto output = m_outputs.at(i).toObject();
-        //        qDebug() << output;
-        const auto script = ParseByteArray(output.value("scriptpubkey"));
-        if (script.size() == 0) {
-            QByteArray data;
-            QDataStream stream(&data, QIODevice::WriteOnly);
-            stream.setByteOrder(QDataStream::BigEndian);
-            stream << uint8_t(0x01);
-            const auto asset_id = ParseByteArray(output.value("asset_id"));
-            Q_ASSERT(asset_id.size() == 32);
-            for (int j = asset_id.size() - 1; j >= 0; --j) {
-                stream << uint8_t(asset_id.at(j));
-            }
-            stream << uint8_t(0x01) << ParseSatoshi(output.value("satoshi"));
-            res.append({{}, data});
-        } else {
-            res.append({{}, m_commitments.at(i).mid(64, 143)});
-        }
-
-        if (output.contains("blinding_key")) {
-            const auto blinding_key = ParseByteArray(output.value("blinding_key"));
-            const auto eph_private_key = ParseByteArray(output.value("eph_private_key"));
-            QByteArray eph_keypair_pub(EC_PUBLIC_KEY_LEN, 0);
-            int rc = wally_ec_public_key_from_private_key(
-                (const unsigned char*) eph_private_key.constData(), eph_private_key.size(),
-                (unsigned char*) eph_keypair_pub.data(), eph_keypair_pub.size());
-            Q_ASSERT(rc == WALLY_OK);
-            res.append({{}, eph_keypair_pub});
-            res.append({{}, blinding_key});
-        } else {
-            QByteArray nonce(1, 0);
-            res.append({{}, nonce});
-            res.append({{}, nonce});
-        }
-
-        {
-            QByteArray data;
-            QDataStream stream(&data, QIODevice::WriteOnly);
-            stream.setByteOrder(QDataStream::LittleEndian);
-            if (script.size() == 0) {
-                stream << varint<uint32_t>(0);
-            } else {
-                stream << varint<uint32_t>(script.size());
-                stream.writeRawData(script.data(), script.size());
-            }
-
-            QJsonObject message = {
-                                   { "index", i },
-                                   { "output", output },
-                                   };
-            res.append({message, data});
-        }
-    }
-    return res;
-}
-
-
-void LedgerGetBlindingFactorsActivity::finalizeLiquidInputFull()
-{
-    int i = 0;
-    m_output_liquid_bytes = outputLiquidBytes();
-    for (const auto& data : m_output_liquid_bytes) {
-        auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_HASH_INPUT_FINALIZE_FULL, i == m_output_liquid_bytes.size()-1 ? 0x80 : 0x00, 0x00, data.second));
-        connect(c, &Command::finished, [this, i] {
-            if (i + 1 < m_output_liquid_bytes.size()) {
-                setMessage(m_output_liquid_bytes.at(i + 1).first);
-            } else {
-                setMessage({});
-            }
-            // qDebug() << "RECEIVE BTCHIP_INS_HASH_INPUT_FINALIZE_FULL" << i << c->m_response.toHex();
-        });
-        connect(c, &Command::error, [] {
-            // qDebug() << "RECEIVE BTCHIP_INS_HASH_INPUT_FINALIZE_FULL" << i << "!!!!! ERROR !!!!!";
-        });
-
-        i++;
-    }
-
-    {
-        QByteArray data(m_inputs.size(), 0);
-        auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_GET_LIQUID_ISSUANCE_INFORMATION, 0x80, 0x00, data));
-        connect(c, &Command::finished, [c] {
-            //            qDebug() << "RECEIVE BTCHIP_INS_GET_LIQUID_ISSUANCE_INFORMATION" << c->m_response.toHex();
-        });
-        connect(c, &Command::error, [i] {
-            //            qDebug() << "RECEIVE BTCHIP_INS_GET_LIQUID_ISSUANCE_INFORMATION" << "!!!!! ERROR !!!!!";
-        });
-    }
-
-    for (i = 0; i < m_outputs.size(); ++i) {
-        const auto commitment = m_commitments.at(i);
-        if (commitment.isEmpty()) {
-            m_asset_commitments.append(QByteArray());
-            m_value_commitments.append(QByteArray());
-        } else {
-            m_asset_commitments.append(commitment.mid(69, 33));
-            m_value_commitments.append(commitment.mid(69 + 33, 33));
-        }
-    }
-
-    const uint32_t locktime = ParseLocktime(m_transaction.value("transaction_locktime"));
-
-    for (i = 0; i < m_hw_inputs.size(); ++i) {
-        const auto input = m_inputs.at(i).toObject();
-        const auto script = ParseByteArray(input.value("prevout_script"));
-
-        startUntrustedTransaction(false, 0, {m_hw_inputs.at(i)}, {m_hw_sequences.at(i)}, script);
-
-        {
-            // untrustedLiquidHashSign
-            QByteArray data;
-            QDataStream stream(&data, QIODevice::WriteOnly);
-
-            QVector<uint32_t> user_path = ParsePath(input.value("user_path"));
-            const auto path = pathToData(user_path);
-
-            stream.writeRawData(path.data(), data.size());
-
-            stream.setByteOrder(QDataStream::BigEndian);
-
-            // TODO android implementation uses empty pin, like:
-            //    stream << uint8_t(0x00);
-            QString pin = "0";
-            auto _pin = pin.toUtf8();
-            stream.writeRawData(path.data(), path.size());
-            stream << uint8_t(pin.size());
-            stream.writeRawData(_pin.data(), _pin.size());
-            stream << uint32_t(locktime);
-            stream << uint8_t(1); // SIGHASH_ALL
-
-            auto c = exchange(apdu(BTCHIP_CLA, BTCHIP_INS_HASH_SIGN, 0x00, 0x00, data));
-            connect(c, &Command::finished, [this, c] {
-                QByteArray signature;
-                signature.append(0x30);
-                signature.append(c->m_response.mid(1));
-                m_sigs.append(signature);
-
-                if (m_sigs.size() == m_hw_inputs.size()) {
-                    m_abfs = m_abfs.mid(m_inputs.size());
-                    m_vbfs = m_vbfs.mid(m_inputs.size());
-                    finish();
-                }
-            });
-        }
-    }
-}
-
-*/