+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed:
 - QR Code for Authenticator app in Two-Factor authentication settings
 - Only show fee amount for redeposits
+- Sign Liquid transaction with Ledger Nano S
 
 ## [2.0.3] - 2024-03-28
 ### Added: